move components to app folder
