+++ src/app/(curriculum)/components/Header/index.tsx
