+++ src/app/(curriculum)/components/Header/locale/en.json
