+++ src/app/(curriculum)/components/Header/locale/pt.json
