+++ src/app/(curriculum)/components/curriculum.module.scss
