+++ src/app/(main)/components/Layout/Background/background.module.scss
