+++ src/app/(main)/components/Layout/Background/index.tsx
