+++ src/app/(main)/components/Layout/Header/sidebar.module.scss
