+++ src/app/(main)/components/Layout/Main/main.module.scss
