+++ src/app/(main)/layout.tsx
@@ -1,4 +1,4 @@
-import Layout from "@/components/Layout";
+import Layout from "./components/Layout";
 
 const MainLayout = ({ children }: { children: React.ReactNode }) => <Layout>{children}</Layout>;
