+++ src/app/(main)/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from "next";
 
-import Presentation from "@/components/pages/home/Presentation";
+import Presentation from "./components/Presentation";
 
 export const metadata: Metadata = {
   title: "Silas Henrique",
