+++ src/app/(main)/resume/components/Content/Header.tsx
@@ -1,4 +1,4 @@
-import ImageModal from "@/components/common/ImageModal";
+import ImageModal from "@/components/ImageModal";
 
 import styles from "./header.module.scss";
