+++ src/app/(main)/resume/components/Content/content.module.scss
