+++ src/app/(main)/resume/components/Dialog/dialog.module.scss
