+++ src/app/(main)/resume/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from "next";
 
-import Content from "@/components/pages/resume/Content";
-import Dialog from "@/components/pages/resume/Dialog";
+import Content from "./components/Content";
+import Dialog from "./components/Dialog";
 
 export const metadata: Metadata = {
   title: "Resume",
